Q14: Rearrange an array in alternating positive and negative numbers
Given an array of positive and negative numbers, arrange them in an
alternate fashion such that every positive number is followed by
negative and vice-versa maintaining the order of appearance. Number of
positive and negative numbers need not be equal. If there are more
positive numbers they appear at the end of the array. If there are more
negative numbers, they too appear at the end of the array.